#50 旅行プラン新規作成時の旅行概要を分割、namedエクスポートに統一
