+++ frontend/src/App.jsx
@@ -1,12 +1,10 @@
 import { Route, Routes } from 'react-router-dom'
 import './styles/App.css'
 
-function App() {
+export function App() {
   return (
     <>
       
     </>
   )
 }
-
-export default App
