+++ frontend/src/components/PageTitle/PageTitle.jsx
@@ -1,0 +1,5 @@
+export function PageTitle({children}) {
+  return (
+    <h2>{children}</h2>
+  );
+}
