+++ frontend/src/components/ProtectedRoute.jsx
@@ -2,7 +2,7 @@
 import { Navigate } from 'react-router-dom';
 import { useAuth } from '@/contexts/AuthContext';
 
-const ProtectedRoute = ({ children }) => {
+export const ProtectedRoute = ({ children }) => {
   const { isAuthenticated, loading } = useAuth();
 
   if (loading) {
@@ -15,5 +15,3 @@ const ProtectedRoute = ({ children }) => {
   
   return children;
 };
-
-export default ProtectedRoute;
