+++ frontend/src/components/button/BackToManagementButton.jsx
@@ -1,0 +1,5 @@
+export function BackToManagementButton({children}) {
+  return (
+    <button>{children}</button>
+  );
+}
