+++ frontend/src/components/button/PlanCreateButton.jsx
@@ -1,0 +1,7 @@
+export function PlanCreateButton({ handleCreate, children }) {
+  return (
+    <button type="button" onClick={handleCreate}>
+      {children}
+    </button>
+  );
+}
