+++ frontend/src/components/planCreate/CurrentDay.jsx
@@ -1,9 +1,7 @@
 import React from "react";
 
-function CurrentDay({ index, onSelectedDayChange }) {
+export function CurrentDay({ index, onSelectedDayChange }) {
   return (
     <button key={index} onClick={() => onSelectedDayChange(index)}>Day {index + 1}</button>
   )
 }
-
-export default CurrentDay;
