+++ frontend/src/main.jsx
@@ -3,7 +3,7 @@ import { Provider } from "./components/ui/provider"
 import { createRoot } from 'react-dom/client'
 import { RouterProvider } from 'react-router-dom'
 
-import routesBasic from './routes/routesBasic.jsx'
+import { routesBasic } from './routes/routesBasic.jsx'
 import { AuthProvider } from './contexts/AuthContext'
 // import './styles/index.css'
