+++ frontend/src/pages/NewTrip.jsx
@@ -1,82 +1,24 @@
-import { useState } from "react";
-import { Link, useNavigate } from "react-router-dom";
-import { parseError, ERROR_MESSAGES } from '@/utils/errorHandler';
-import { planOverviewCreate } from "@/api/planOverviewApi";
+import { useNewTrip } from "@/hooks/tripPlan/useNewTrip";
+import { NewTripOverview } from "@/components/planCreate/NewTripOverview";
+import { PlanCreateButton } from "@/components/button/PlanCreateButton";
+import { BackToManagementButton } from "@/components/button/BackToManagementButton";
+import { PageTitle } from "@/components/PageTitle/PageTitle";
+import { Link } from "react-router-dom";
 
 // 旅行名と日付を入力するページ
-export default function NewTrip() {
-  const [tripData, setTripData] = useState({});
-  const navigate = useNavigate();
-  const [error, setError] = useState('');
-
-  const handleTrip = (e) => {
-    setTripData({
-      ...tripData,
-      [e.target.name]: e.target.value,
-    });
-    // エラーメッセージをクリア
-    setError('');
-  };
-
-  // API通信
-  const handleCreate = async () => {
-    try {
-      const response = await planOverviewCreate(tripData);
-      
-      setError('');
-      // プランIDのみを渡してTripPlanページに遷移
-      navigate(`/trip-plan/${response.data.id}`);
-    } catch (error) {
-      const { message } = parseError(error, ERROR_MESSAGES.PLAN_CREATE_FAILED);
-      setError(message);
-    }
-  };
+export function NewTripPage() {
+  const { tripData, error, handleTrip, handleCreate } = useNewTrip();
 
   return (
     <div>
-      <h2>新しい旅行プランを作成</h2>
+      <PageTitle>旅行プラン作成</PageTitle>
       <Link to="/management">
-          <button>管理画面へ戻る</button>
+        <BackToManagementButton>
+          管理画面へ戻る
+        </BackToManagementButton>
       </Link>
-      <div>
-        <p>{error}</p>
-        <div>
-          <label htmlFor="tripTitle">旅行タイトル</label>
-          <input
-            type="text"
-            id="tripTitle"
-            name="tripTitle"
-            value={tripData.tripTitle || ''}
-            placeholder="旅行タイトル"
-            onChange={handleTrip}
-          />
-        </div>
-        <div>
-          <label htmlFor="startDate">出発日</label>
-          <input
-            type="date"
-            id="startDate"
-            name="startDate"
-            value={tripData.startDate || ''}
-            placeholder="出発日"
-            onChange={handleTrip}
-          />
-        </div>
-        <div>
-          <label htmlFor="endDate">帰着日</label>
-          <input
-            type="date"
-            id="endDate"
-            name="endDate"
-            value={tripData.endDate || ''}
-            placeholder="帰着日"
-            onChange={handleTrip}
-          />
-        </div>
-      </div>
-      <button type="button" onClick={handleCreate}>
-        作成
-      </button>
+      <NewTripOverview tripData={tripData} error={error} handleTrip={handleTrip} />
+      <PlanCreateButton handleCreate={handleCreate}>作成</PlanCreateButton>
     </div>
   );
 }
