+++ frontend/src/routes/routesBasic.jsx
@@ -1,14 +1,14 @@
 import { Route, createBrowserRouter, createRoutesFromElements } from "react-router-dom";
-import Home from "@/pages/Home";
-import Management from "@/pages/Management";
-import TripPlan from "@/pages/TripPlan";
-import NewTrip from "@/pages/NewTrip";
-import Login from "@/pages/Login";
-import Register from "@/pages/Register";
-import ProtectedRoute from "@/components/ProtectedRoute";
+import { Home } from "@/pages/Home";
+import { Management } from "@/pages/Management";
+import { TripPlan } from "@/pages/TripPlan";
+import { NewTripPage } from "@/pages/NewTrip";
+import { Login } from "@/pages/Login";
+import { Register } from "@/pages/Register";
+import { ProtectedRoute } from "@/components/ProtectedRoute";
 
 
-const routesBasic = createBrowserRouter(
+export const routesBasic = createBrowserRouter(
   createRoutesFromElements(
     <>
       <Route path="/" element={<Home />} />
@@ -24,7 +24,7 @@ const routesBasic = createBrowserRouter(
         }/>
         <Route path="/new-trip" element={
           <ProtectedRoute>
-            <NewTrip />
+            <NewTripPage />
           </ProtectedRoute>
         }/>
       <Route path="/login" element={<Login />} />
@@ -32,5 +32,3 @@ const routesBasic = createBrowserRouter(
     </>
   )
 );
-
-export default routesBasic;
